refactor of insert-data endpoint, exceptions added

+++ src/main/java/org/saebio/requesthandler/backup/DeleteBackupHandler.java
@@ -1,13 +1,14 @@
 package org.saebio.requesthandler.backup;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonElement;
 import org.saebio.api.Answer;
 import org.saebio.api.HttpStatus;
 import org.saebio.backup.Backup;
 import org.saebio.backup.BackupApiConstants;
 import org.saebio.backup.BackupService;
 import org.saebio.requesthandler.AbstractRequestHandler;
+import org.saebio.requesthandler.exception.AbstractRequestException;
+import org.saebio.requesthandler.exception.InvalidRequestBodyObjectException;
+import org.saebio.requesthandler.exception.RequestBodyObjectNotFoundException;
 
 import java.util.Map;
 
@@ -18,9 +19,13 @@ public class DeleteBackupHandler extends AbstractRequestHandler<Backup> {
     }
 
     @Override
-    protected Answer processImpl(Backup value, Map<String, String> queryParams) {
+    protected Answer processImpl(Backup value, Map<String, String> queryParams) throws AbstractRequestException {
+        if (!value.isValid()) {
+            throw new InvalidRequestBodyObjectException();
+        }
+
         if (!BackupService.backupExists(value)) {
-            return new Answer(BackupApiConstants.ERROR_BACKUP_DOES_NOT_EXIST, HttpStatus.BadRequest());
+            throw new RequestBodyObjectNotFoundException();
         }
 
         if (BackupService.removeBackup(value)) {
